refactor(app): extract macro orchestrators to macro_actions (phase 3b)
Move the macro record/playback orchestrators out of render.rs into a new
app/macro_actions.rs sibling. Two non-contiguous blocks:

  3082-3165 (84 lines): toggle_macro_recording, start_macro_recording,
                        build_macro_stop_hint, stop_macro_recording,
                        build_macro_play_hint
  3236-3442 (207 lines): play_macro, record_macro_action,
                         show_macro_in_buffer, list_macros_in_buffer

`recompute_layout` (3171-3234) stays in render.rs — it's layout code
that happens to be called by play_macro but uses the same ratatui types
and layout primitives the rest of render.rs depends on.

These methods were already going through the MacroState API extracted
in phase 2a; this commit is purely structural — moves their definitions
into a topic-specific file with no behaviour change.

render.rs goes from 3893 -> 3602 lines (-291). 2026 tests still pass.
